Enhance Makefile Dependency Management and Update Lyn Event Files
- Improved dependency flags in the Makefile to ensure unique naming for dependency files, preventing collisions during builds.
- Updated various Lyn event files to correct word entries and ensure consistency in data structure.
- Adjusted the designer configuration to enable new features and improve overall functionality across multiple components.
- Enhanced the Usability and StatScreen event files to refine command availability and improve user experience during gameplay.

+++ Kernel/Wizardry/Misc/VeslyAchievements/AchievementsMenu.c
@@ -274,7 +274,7 @@ const char * GetStringNextLine_Achievements(const char * str, int maxWidth)
                 nextWordWidth += gw;
             }
 
-            if (ttlWidth + width + nextWordWidth > maxWidth)
+            if ((int)(ttlWidth + width + nextWordWidth) > maxWidth)
             {
                 // Break here before the space and return pointer to first character of next word
                 const char * nextWordStart = charEnd; // skip the space itself
@@ -284,7 +284,7 @@ const char * GetStringNextLine_Achievements(const char * str, int maxWidth)
             }
         }
 
-        if (ttlWidth + width > maxWidth)
+        if ((int)(ttlWidth + width) > maxWidth)
         {
             // Break at this character
             return iter;
